Do not expose full path in error message
If you modify the url parameters when in Manager, you can currently produce an error message that exposes the full path of the application.  Whilst this is low risk as it only applies to already authenticated users, it would be preferable if the full path was not shown on the page, but only in the error log which is only visible to users with appropriate permissions.  This commit removes the file path from the on-page error message.